+++ Runtime/EventSystem/GameEventBool.cs
@@ -5,7 +5,7 @@ namespace PushForward.EventSystem
 
     /// <summary>An extension of the game event that contains a boolean.</summary>
     [CreateAssetMenu(menuName = "ScriptableObjects/Event System/Game Event Bool", order = 22)]
-    class GameEventBool : GameEvent
+    public class GameEventBool : GameEvent
     {
         public bool condition;